+++ platform/datacollector/src/routes/health.rs
@@ -1,0 +1,6 @@
+use actix_web::{get, HttpResponse};
+
+#[get("/health")]
+pub async fn handler() -> HttpResponse {
+    HttpResponse::Ok().body("OK")
+}
